Renaming file to be compliant with the StyleCop Analyzer
