+++ libraries/Microsoft.Bot.Builder.Ai.LUIS/Generator/NumberWithUnits.cs
@@ -37,56 +37,4 @@ namespace Microsoft.Bot.Builder.Ai.LUIS
         [JsonProperty("units")]
         public string Units { get; }
     }
-
-    /// <summary>
-    /// Strongly typed LUIS builtin_age.
-    /// </summary>
-    public class Age : NumberWithUnits
-    {
-        public Age(double number, string units)
-            : base(number, units)
-        {
-        }
-
-        public override string ToString() => $"Age({Number} {Units})";
-    }
-
-    /// <summary>
-    /// Strongly typed LUIS builtin_dimension.
-    /// </summary>
-    public class Dimension : NumberWithUnits
-    {
-        public Dimension(double number, string units)
-            : base(number, units)
-        {
-        }
-
-        public override string ToString() => $"Dimension({Number} {Units})";
-    }
-
-    /// <summary>
-    /// Strongly typed LUIS builtin_money.
-    /// </summary>
-    public class Money : NumberWithUnits
-    {
-        public Money(double number, string units)
-            : base(number, units)
-        {
-        }
-
-        public override string ToString() => $"Currency({Number} {Units})";
-    }
-
-    /// <summary>
-    /// Strongly typed LUIS builtin_temperature.
-    /// </summary>
-    public class Temperature : NumberWithUnits
-    {
-        public Temperature(double number, string units)
-            : base(number, units)
-        {
-        }
-
-        public override string ToString() => $"Temperature({Number} {Units})";
-    }
 }
